requirements.txt
Django>=4.2
djangorestframework>=3.14
django-oauth-toolkit>=3.0.1
requests>=2.31
psycopg2-binary>=2.9
django-cors-headers>=3.14
gunicorn>=20.1.0